lib/api_evalution/model_Evalution/requirements.txt
fastapi==0.115.11
uvicorn==0.21.1
python-dotenv==1.0.1
pydantic==2.10.6
torch==2.6.0
transformers==4.47.0
datasets==3.1.0
evaluate==0.4.3
accelerate==1.1.1
bitsandbytes==0.44.1